+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a public Python API `gantry.api` with some useful methods.
 
+### Fixed
+
+- Added work-around for when `ensurepip` package is not installed.
+
 ## [v2.3.0](https://github.com/allenai/beaker-gantry/releases/tag/v2.3.0) - 2025-05-15
 
 ### Added